// src/server/api/generated/schemas/jobPostings.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * # GetJob AI

Welcome to the **GetJob AI** API documentation.

 * OpenAPI spec version: 14.11
 */

/**
 * Collection of user job postings.
 */
export interface JobPostings {
  /** Unique identifier for the job posting.

  Note:
  This is a Primary Key.<pk/> */
  id: string;
  /** ID of the job posting owner. */
  user_id: string;
  /** Structured job posting data in JSON format. */
  content: unknown;
  created_at: string;
  updated_at: string;
}